chore: fix compile error

+++ taos-ws/src/lib.rs
@@ -551,7 +551,7 @@ impl TaosBuilder {
                         self.to_ws_url(),
                         retries
                     );
-                    tokio::time::sleep(std::time::Duration::from_secs(retries * 500)).await;
+                    tokio::time::sleep(std::time::Duration::from_secs(retries as u64 * 500)).await;
                 }
             }
         }
